requirements.txt
rule_engine==2.4.1
pyyaml==5.4.1
exchangelib==3.3.2